requirements.txt
yfinance
pandas==2.2.0
ta==0.11.0
litellm==1.28.1
feedparser==6.0.11
python-dotenv==1.0.1
google-generativeai
alpaca-py